moved retrival of likes to separate module, front end reflects retrieving likes
