+++ index/src/css/index.css
@@ -1,7 +1,1 @@
-.headerbox {
-    font-size: 1.5em;
-    letter-spacing: .3em;
-    position:absolute;top:35%;
-    left:50%;transform:translate(-50%,-50%);
-    -ms-transform:translate(-50%,-50%)
- }
+
