+++ index/src/css/login.css
@@ -10,6 +10,14 @@
 	width: 40%;
 }
 
+.headerbox {
+    font-size: 1.5em;
+    letter-spacing: .3em;
+    position:absolute;top:35%;
+    left:50%;transform:translate(-50%,-50%);
+    -ms-transform:translate(-50%,-50%)
+ }
+
 #small {
 	font-size: .9em;
 	letter-spacing: 0;
